revert 'update saml.ts'
